+++ package.json
@@ -13,12 +13,13 @@
     "gen:grammar:old": "ohm generateBundles --withTypes src/050-grammar/prev/*.ohm",
     "gen:grammar:new": "pgen src/050-grammar/next/grammar.gg -o src/050-grammar/next/grammar.ts",
     "gen:grammar": "yarn gen:grammar:old && yarn gen:grammar:new",
-    "gen:pack": "ts-node ./scripts/pack.ts",
+    "gen:pack": "ts-node src/090-func/pack.build.ts",
     "gen:compiler": "ts-node ./scripts/prepare.ts",
     "gen": "yarn gen:grammar && yarn gen:pack && yarn gen:compiler",
     "clean": "rm -fr dist",
     "cleanall": "rm -fr dist node_modules",
-    "build": "tsc && node --no-warnings=ExperimentalWarning -r ts-node/register ./scripts/copy-files",
+    "build:func": "ts-node ./src/090-func/copy.build.ts",
+    "build": "tsc && yarn build:func && yarn node --no-warnings=ExperimentalWarning -r ts-node/register ./scripts/copy-files",
     "test": "yarn gen:grammar && jest",
     "coverage": "cross-env COVERAGE=true jest",
     "release": "yarn clean && yarn build && yarn coverage && yarn release-it --npm.yarn1",
