+++ src/090-func/.gitignore
@@ -1,0 +1,1 @@
+funcfiftlib.wasm.js
