+++ src/090-func/__testdata__/small.fc
