+++ src/090-func/funcfiftlib.js
